Localisation view refactor (#1474)
