+++ nusight2/src/client/components/localisation/r3f_components/bounding_box/wall.vert
@@ -1,0 +1,5 @@
+varying vec3 vPosition;
+void main() {
+    vPosition   = position;
+    gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.);
+}
