+++ nusight2/src/client/components/localisation/r3f_components/field_line_points/view.tsx
@@ -1,0 +1,24 @@
+import React from "react";
+
+import { Vector3 } from "../../../../../shared/math/vector3";
+import { LocalisationModel } from "../../model";
+
+export const FieldLinePoints = ({ model }: { model: LocalisationModel }) => (
+  <>
+    {model.robots.map(
+      (robot) =>
+        robot.visible && (
+          <object3D key={robot.id}>
+            {robot.rPFf.map((d, i) => {
+              return (
+                <mesh key={String(i)} position={d.add(new Vector3(0, 0, 0.005)).toArray()}>
+                  <circleBufferGeometry args={[0.02, 20]} />
+                  <meshBasicMaterial color="blue" />
+                </mesh>
+              );
+            })}
+          </object3D>
+        ),
+    )}
+  </>
+);
