+++ nusight2/src/client/components/localisation/r3f_components/localised_goals/view.tsx
@@ -1,0 +1,29 @@
+import React from "react";
+
+import { Vector3 } from "../../../../../shared/math/vector3";
+import { LocalisationModel } from "../../model";
+
+export const Goals = ({ model }: { model: LocalisationModel }) => (
+  <>
+    {model.robots.map(
+      (robot) =>
+        robot.visible &&
+        robot.rGFf && (
+          <object3D key={robot.id}>
+            {robot.rGFf.map((goal, i) => {
+              return (
+                <mesh
+                  key={String(i)}
+                  position={goal.bottom.add(new Vector3(0, 0, goal.top.z / 2)).toArray()}
+                  rotation={[Math.PI / 2, 0, 0]}
+                >
+                  <cylinderBufferGeometry args={[0.05, 0.05, goal.top.z, 20]} />
+                  <meshStandardMaterial color="magenta" />
+                </mesh>
+              );
+            })}
+          </object3D>
+        ),
+    )}
+  </>
+);
