+++ nusight2/src/client/components/localisation/r3f_components/purpose_label/view.tsx
@@ -1,0 +1,28 @@
+import React from "react";
+
+import { LocalisationRobotModel } from "../../robot_model";
+import { TextBillboard } from "../text_billboard/view";
+
+export const PurposeLabel = ({
+  robotModel,
+  cameraPitch,
+  cameraYaw,
+}: {
+  robotModel: LocalisationRobotModel;
+  cameraPitch: number;
+  cameraYaw: number;
+}) => {
+  const rTFf = robotModel.Hft.decompose().translation;
+  const label = robotModel.player_id == -1 ? robotModel.purpose : "N" + robotModel.player_id + " " + robotModel.purpose;
+
+  return (
+    <TextBillboard
+      position={[rTFf?.x, rTFf?.y, rTFf?.z + 0.6]}
+      backgroundColor={robotModel.color}
+      textColor="white"
+      text={label}
+      cameraPitch={cameraPitch}
+      cameraYaw={cameraYaw}
+    />
+  );
+};
