+++ nusight2/src/client/components/localisation/r3f_components/skybox/skybox.vert
