+++ nusight2/src/client/components/localisation/r3f_components/text_geometry_helper/TextGeometryHelper.tsx
@@ -1,0 +1,13 @@
+import { TextGeometry } from "three/examples/jsm/geometries/TextGeometry";
+import { FontLoader } from "three/examples/jsm/loaders/FontLoader";
+
+import roboto from "./Roboto Medium_Regular.json";
+
+export const TextGeometryHelper = (x: string) => {
+  const font = new FontLoader().parse(roboto);
+  return new TextGeometry(x, {
+    font: font,
+    size: 0.1,
+    height: 0,
+  }).center();
+};
