--- nusight2/src/client/components/localisation/style.module.css
@@ -1,117 +1,0 @@
-.localisation {
-  flex-grow: 1;
-  flex-shrink: 1;
-  display: flex;
-  flex-direction: column;
-  position: relative;
-}
-
-.localisation__menu {
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-  height: 100%;
-  display: flex;
-  align-items: stretch;
-}
-
-.localisation__menuItem {
-  display: flex;
-  margin: 0;
-  padding: 0;
-}
-
-.localisation__menuButton {
-  align-items: center;
-  background-color: transparent;
-  border: 0;
-  cursor: pointer;
-  display: flex;
-  flex-direction: column;
-  font-size: 0.8rem;
-  height: 60px;
-  width: 128px;
-  justify-content: center;
-  outline: 0;
-  padding: 0 1em;
-  transition: background-color 0.3s ease;
-}
-
-.localisation__menuButton:hover {
-  background-color: #f9f9f9;
-  border: 1px solid rgba(0, 0, 0, 0.2);
-}
-
-.localisation__menuButton:focus,
-.localisation__menuButton:active {
-  outline: none;
-  border: 1px solid rgba(0, 0, 0, 0.4);
-}
-
-.localisation__canvas {
-  flex-grow: 1;
-  position: relative;
-}
-
-.localisation__status {
-  background-color: rgba(0, 0, 0, 0.3);
-  border-radius: 6px;
-  color: white;
-  padding: 16px;
-  text-align: center;
-  position: absolute;
-  bottom: 8px;
-  left: 8px;
-  right: 8px;
-  font-size: 1.1rem;
-  font-weight: bold;
-  display: flex;
-  justify-content: space-between;
-  flex-basis: 0;
-}
-
-.localisation__info {
-  text-align: left;
-  width: 33%;
-}
-
-.localisation__target {
-  width: 33%;
-}
-
-.localisation__viewMode {
-  text-align: right;
-  width: 33%;
-}
-
-.fieldOption {
-  border-bottom-left-radius: 2px;
-  border-bottom-right-radius: 2px;
-  overflow: hidden;
-  width: 128px;
-}
-
-.fieldOption {
-  box-sizing: border-box;
-  display: flex;
-  min-width: 104px;
-}
-
-.fieldOption:hover {
-  background-color: #eee;
-}
-
-.fieldOption {
-  display: flex;
-  align-items: center;
-  flex-grow: 1;
-  height: 40px;
-  padding: 0;
-  padding-left: 1em;
-  cursor: pointer;
-}
-
-.checkmark {
-  margin-left: 10px;
-  color: green; /* Green checkmark for visual cue */
-}
